requirements.txt
pillow
pytesseract
python-docx
pdf2docx
docx2pdf
